#53 fix : 작성자만 삭제 가능하도록 수정

+++ src/main/java/com/example/myrok/service/RecordServiceImpl.java
@@ -86,20 +86,21 @@ public class RecordServiceImpl implements RecordService{
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public void deleteUpdate(Long recordId) {
+    public void deleteUpdate(Long recordId, Long recordWriterId) {
         // 회의록 삭제
         Record record = recordRepository.findById(recordId)
                 .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 회의록입니다. id: " + recordId));
+        // 이미 삭제된 회의록이면 예외
         if (record.getDeleted()) {
             throw new CustomException(ErrorCode.DELETED_RECORD_CODE, HttpStatus.BAD_REQUEST);
         }
-//        //회의 참여자가 아니라면 예외
-//        MemberRecord memberRecord =memberRecordRepository.findByMemberIdAndRecordIdAndDeletedIsFalse(recordWriterId,recordId)
-//                .orElseThrow(() -> new CustomException(ErrorCode.WRONG_RECORD_ACCESS, HttpStatus.BAD_REQUEST));
-//        //작성자 외 수정 시도시 예외
-//        if (memberRecord.getRole()!=Role.ADMIN){
-//            throw new CustomException(ErrorCode.WRONG_UPDATE_ACCESS, HttpStatus.BAD_REQUEST);
-//        }
+        //회의 참여자가 아니라면 예외
+        MemberRecord memberRecord =memberRecordRepository.findByMemberIdAndRecordIdAndDeletedIsFalse(recordWriterId,recordId)
+                .orElseThrow(() -> new CustomException(ErrorCode.WRONG_RECORD_ACCESS, HttpStatus.BAD_REQUEST));
+        //작성자 외 삭제 시도시 예외
+        if (memberRecord.getRole()!=Role.ADMIN){
+            throw new CustomException(ErrorCode.WRONG_DELETE_ACCESS, HttpStatus.BAD_REQUEST);
+        }
         record.delete();
         recordRepository.save(record);
